visionlabs_implement

+++ image.py
@@ -14,8 +14,8 @@ def load_data_fidt(img_path, args, train=True):
     while True:
         try:
             gt_file = h5py.File(gt_path)
-            k = np.asarray(gt_file['kpoint'])
-            fidt_map = np.asarray(gt_file['fidt_map'])
+            k = np.asarray(gt_file['kpoint']).astype(np.float32, copy=False)
+            fidt_map = np.asarray(gt_file['fidt_map']).astype(np.float32, copy=False)
             break
         except OSError:
             print("path is wrong, can not load ", img_path)
